+++ src/routes/ViewAllLogs/ViewAllLogs.test.js
@@ -186,6 +186,7 @@ const renderViewAllLogs = query => {
             pathname: '/job-logs',
             search: '?sort=-completedDate',
           }}
+          match={{ params: {} }}
           refreshRemote={noop}
         />
       </ModuleHierarchyProvider>