+++ src/settings/FileExtensions/index.js
@@ -1,1 +1,4 @@
 export * from './FileExtensions';
+export * from './CreateFileExtension';
+export * from './EditFileExtension';
+export * from './ViewFileExtension';